requirements.txt
fastapi
uvicorn
pydantic
sqlmodel
alembic
psycopg2-binary
asyncpg # Added for async PostgreSQL driver
celery
redis
python-multipart
opencv-python
pdf2image
Pillow
transformers
torch
accelerate
bitsandbytes
scikit-learn
matplotlib
seaborn
jupyter
notebook
structlog
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-celery
prometheus_client
locust
ruff
black
mypy
pytest
pytest-cov